:bug: navigation: fix backUntil
Motivation: `Get.backUntil` or, more specific,  `GetRouterDelegate.backUntil` checks if the number of `_activePages.length <= 1`, which does not make sense. `GetRouterDelegate.offUntil`, which has similar semantics, tests if `_activePages.isNotEmpty`. So in order to make `backUntil` work as expected and to be consistent with `offUntil` I propose the following change.